# turbo-mailer-server/docs/swagger.yaml
basePath: /
definitions:
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.Pool:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      sender_count:
        type: integer
      senders:
        items:
          $ref: '#/definitions/models.PoolSender'
        type: array
      updated_at:
        type: string
    type: object
  models.PoolSender:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      domain:
        type: string
      from_email:
        type: string
      from_name:
        type: string
      id:
        type: integer
      pool_id:
        type: integer
      reply_to:
        type: string
      updated_at:
        type: string
    required:
    - domain
    - from_email
    - from_name
    type: object
  models.StringArray:
    properties:
      v:
        items:
          type: string
        type: array
    type: object
  models.Task:
    properties:
      content:
        type: string
      content_type:
        description: MIME type
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      last_dispatch_at:
        type: string
      max_dispatch_pre_hour:
        type: integer
      metadata:
        items:
          type: integer
        type: array
      pools:
        items:
          $ref: '#/definitions/models.TaskPool'
        type: array
      receivers:
        $ref: '#/definitions/models.StringArray'
      schedule_at:
        type: string
      state:
        type: string
      subject:
        type: string
      updated_at:
        type: string
    type: object
  models.TaskPool:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      pool:
        $ref: '#/definitions/models.Pool'
      pool_id:
        type: integer
      task_id:
        type: integer
      updated_at:
        type: string
      weight:
        type: integer
    type: object
...
  schema.Page-models_Task:
    properties:
      currentPage:
        description: Current page number
        type: integer
      list:
        description: List of data items
        items:
          $ref: '#/definitions/models.Task'
        type: array
      next:
        description: URL for the next page
        type: string
      prev:
        description: URL for the previous page
        type: string
      total:
        description: Total number of records
        type: integer
      totalPages:
        description: Total number of pages
        type: integer
    type: object
info:
  contact: {}
  title: Turbo Mailer API
  version: 0.1.0
paths:
...
  /api/v1/tasks:
    get:
      consumes:
      - application/json
      description: Get a paginated list of tasks with optional sorting and filtering
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: pageSize
        type: integer
      - default: desc
        description: 'Sort order: ''asc'' or ''desc'''
        in: query
        name: sort
        type: string
      - description: Search by task subject
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.Page-models_Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWT: []
      summary: List all tasks with pagination
      tags:
      - Tasks
    post:
      consumes:
      - multipart/form-data
      description: Create a new task with the provided details and uploaded files
      parameters:
      - description: Task subject
        in: formData
        name: subject
        required: true
        type: string
      - description: Context type (html or text)
        in: formData
        name: context_type
        required: true
        type: string
      - description: Content file (html or text)
        in: formData
        name: content
        required: true
        type: file
      - description: Receivers CSV file
        in: formData
        name: receivers
        required: true
        type: file
      - description: Max dispatch per hour
        in: formData
        name: max_dispatch_per_hour
        type: integer
      - description: 'Schedule at, format: YYYY-MM-DD HH:MM:SS'
        in: formData
        name: schedule_at
        type: string
      - collectionFormat: csv
        description: Pools
        in: formData
        items:
          type: integer
        name: pools
        type: array
      - collectionFormat: csv
        description: Pools weights
        in: formData
        items:
          type: integer
        name: pools_weights
        type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWT: []
      summary: Create a new task
      tags:
      - Tasks
  /api/v1/tasks/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a task by ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWT: []
      summary: Delete a task
      tags:
      - Tasks
    get:
      consumes:
      - application/json
      description: Get details of a specific task by ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWT: []
      summary: Get a specific task
      tags:
      - Tasks
    put:
      consumes:
      - multipart/form-data
      description: Update an existing task's details and files
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Task subject
        in: formData
        name: subject
        type: string
      - description: Context type (html or text)
        in: formData
        name: contextType
        type: string
      - description: Content file (html or text)
        in: formData
        name: content
        type: file
      - description: Receivers CSV file
        in: formData
        name: receivers
        type: file
      - description: Task state
        in: formData
        name: state
        type: string
      - description: Max dispatch per hour
        in: formData
        name: maxDispatchPerHour
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWT: []
      summary: Update a task
      tags:
      - Tasks
  /api/v1/tasks/{id}/start-immediately:
    post:
      consumes:
      - application/json
      description: Immediately start a task by setting its schedule time to now
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWT: []
      summary: Start a task immediately
      tags:
      - Tasks
  /api/v1/tasks/{id}/test:
    post:
      consumes:
      - application/json
      description: Test a task by sending a test email to the specified address
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Email address to send the test to
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWT: []
      summary: Test a task
      tags:
      - Tasks
  /api/v1/tasks/content-template:
    get:
      consumes:
      - application/json
      description: Download a template file for the task content (HTML or plain text)
      parameters:
      - description: Template type (html or text)
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Template file
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWT: []
      summary: Download content template
      tags:
      - Tasks
  /api/v1/tasks/receivers-template:
    get:
      consumes:
      - application/json
      description: Download a template CSV file for the task receivers with example
        email addresses
      produces:
      - text/csv
      responses:
        "200":
          description: CSV template file
          schema:
            type: file
      security:
      - JWT: []
      summary: Download receivers template
      tags:
      - Tasks
swagger: "2.0"